# docs/openapi/zj-integration.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SGA360 - API de integración Z&J",
    "version": "1.1.0",
    "summary": "Integración empresarial para consultar terceros, procesos y órdenes de cliente.",
    "description": "API máquina-a-máquina para los desarrollos externos de Z&J. El cliente obtiene un JWT con vigencia de 3 horas, verifica disponibilidad mediante el handshake, consulta catálogos internos y crea instancias y órdenes de cliente de forma idempotente.\n\n## Flujo recomendado\n\n1. Obtener un JWT en `POST /auth/token`.\n2. Ejecutar `GET /handshake`.\n3. Consultar o buscar terceros, servicios y procesos.\n4. Crear una instancia de proceso.\n5. Crear la orden de cliente asociada.\n\nEl `company_id` y el usuario técnico provienen siempre del JWT. Nunca deben enviarse en el body, query string ni parámetros de ruta.",
    "contact": {
      "name": "SGA360"
    }
  },
  "servers": [
    {
      "url": "/api/integrations/v1",
      "description": "Servidor actual"
    }
  ],
...
  "paths": {
...
    "/third-parties/search": {
      "get": {
        "tags": ["Terceros"],
        "summary": "Buscar terceros",
        "description": "Busca clientes activos de la compañía autenticada.\n\n## Cómo usar los parámetros\n\nTodos los parámetros son opcionales individualmente, pero la petición debe conservar **al menos uno** de estos filtros: `names`, `lastNames`, `corporative_name`, `email` o `nit`. Para cambiar la búsqueda, agregue o quite parámetros del query string. Si envía varios filtros, el tercero debe cumplirlos todos.\n\n- `names`, `lastNames` y `corporative_name`: coincidencia parcial, sin distinguir mayúsculas.\n- `email`: coincidencia exacta, sin distinguir mayúsculas.\n- `nit`: coincidencia exacta; puede enviarse con o sin puntos y guiones.\n- `type`, `page` y `limit`: parámetros adicionales opcionales.\n\n### Ejemplos rápidos\n\nSolo razón social: `?corporative_name=Z%26J`\n\nSolo NIT: `?nit=900.123.456-7`\n\nNombre y apellido: `?names=Juan&lastNames=Pérez`\n\nNombre con opciones: `?names=Juan&type=client,both&page=1&limit=20`\n\nPara personas jurídicas, `corporative_name` consulta el campo interno `names`. Una petición sin ninguno de los cinco filtros de identidad responde con HTTP `400`.",
        "operationId": "searchThirdParties",
        "x-codeSamples": [
          {
            "lang": "Shell",
            "label": "Solo razón social",
            "source": "curl --get \\\n  'http://localhost:3000/api/integrations/v1/third-parties/search' \\\n  --header 'Authorization: Bearer <jwt>' \\\n  --data-urlencode 'corporative_name=Z&J'"
          },
          {
            "lang": "Shell",
            "label": "Solo correo",
            "source": "curl --get \\\n  'http://localhost:3000/api/integrations/v1/third-parties/search' \\\n  --header 'Authorization: Bearer <jwt>' \\\n  --data-urlencode 'email=cliente@empresa.com'"
          },
          {
            "lang": "Shell",
            "label": "Solo NIT",
            "source": "curl --get \\\n  'http://localhost:3000/api/integrations/v1/third-parties/search' \\\n  --header 'Authorization: Bearer <jwt>' \\\n  --data-urlencode 'nit=900.123.456-7'"
          },
          {
            "lang": "Shell",
            "label": "Combinar y quitar parámetros",
            "source": "curl --get \\\n  'http://localhost:3000/api/integrations/v1/third-parties/search' \\\n  --header 'Authorization: Bearer <jwt>' \\\n  --data-urlencode 'names=Juan' \\\n  --data-urlencode 'lastNames=Pérez' \\\n  --data-urlencode 'type=client,both' \\\n  --data-urlencode 'page=1' \\\n  --data-urlencode 'limit=20'\n\n# Para ampliar la búsqueda, quite por ejemplo esta línea:\n#   --data-urlencode 'lastNames=Pérez'\n# Debe dejar al menos names, lastNames, corporative_name, email o nit."
          }
        ],
        "parameters": [
          {
            "name": "names",
            "in": "query",
            "description": "Nombre o fragmento del nombre.",
            "schema": { "type": "string", "maxLength": 200 },
            "example": "Juan"
          },
          {
            "name": "lastNames",
            "in": "query",
            "description": "Apellido o fragmento del apellido.",
            "schema": { "type": "string", "maxLength": 200 },
            "example": "Pérez"
          },
          {
            "name": "corporative_name",
            "in": "query",
            "description": "Razón social o fragmento de ella.",
            "schema": { "type": "string", "maxLength": 200 },
            "example": "Z&J"
          },
          {
            "name": "email",
            "in": "query",
            "description": "Correo electrónico exacto, sin distinguir mayúsculas.",
            "schema": { "type": "string", "format": "email", "maxLength": 200 },
            "example": "cliente@empresa.com"
          },
          {
            "name": "nit",
            "in": "query",
            "description": "NIT o número de identificación exacto. Se ignoran separadores como puntos y guiones.",
            "schema": { "type": "string", "maxLength": 200 },
            "example": "900.123.456-7"
          },
          {
            "name": "type",
            "in": "query",
            "description": "Uno o varios tipos separados por coma. Por defecto: client,both.",
            "schema": { "type": "string", "example": "client,both" }
          },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Limit" }
        ],
        "responses": {
          "200": {
            "description": "Coincidencias paginadas.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/SuccessResponse" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/ThirdParty" }
                        },
                        "pagination": { "$ref": "#/components/schemas/Pagination" }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/third-parties/{id}": {
      "get": {
        "tags": ["Terceros"],
        "summary": "Consultar tercero por ID",
        "operationId": "getThirdPartyById",
        "parameters": [
          {
            "$ref": "#/components/parameters/ResourceId"
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle del tercero.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ThirdParty"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/process-instances/{id}": {
      "get": {
        "tags": ["Procesos"],
        "summary": "Consultar instancia de proceso",
        "operationId": "getProcessInstanceById",
        "parameters": [
          {
            "$ref": "#/components/parameters/ResourceId"
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle de la instancia.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ProcessInstance"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/process-instances": {
      "post": {
        "tags": ["Procesos"],
        "summary": "Crear instancia de proceso",
        "description": "Crea solamente la instancia. La compañía y el responsable salen del JWT; la primera etapa se obtiene de la configuración del proceso. La operación es idempotente por `external_reference`.",
        "operationId": "createProcessInstance",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProcessInstanceRequest"
              },
              "example": {
                "external_reference": "WA-PROCESS-84521",
                "process_id": 1,
                "thirdParty_id": 93,
                "start_date": "2026-07-30T14:15:00-05:00",
                "delivery_date": "2026-07-31T18:00:00-05:00",
                "description": "Pedido recibido mediante WhatsApp",
                "metadata": {
                  "source": "whatsapp",
                  "conversation_id": "573001234567"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Instancia creada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateProcessInstanceResponse"
                }
              }
            }
          },
          "200": {
            "description": "Referencia repetida con el mismo contenido. Se devuelve la instancia existente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateProcessInstanceResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT obtenido mediante `POST /auth/token`."
      }
    },
    "parameters": {
      "ResourceId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "minimum": 1
        }
      },
...
      "Page": {
        "name": "page",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 50
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "La petición no cumple el contrato.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "Unauthorized": {
        "description": "JWT ausente, inválido o vencido.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": {
                "code": "INVALID_ACCESS_TOKEN",
                "message": "Token de acceso inválido o vencido."
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "La credencial no posee el scope requerido.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": {
                "code": "INSUFFICIENT_SCOPE",
                "message": "La integración no tiene permisos para esta operación."
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "El recurso no existe dentro de la compañía autenticada.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Conflict": {
        "description": "La referencia externa ya existe con información diferente.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": {
                "code": "EXTERNAL_REFERENCE_CONFLICT",
                "message": "La referencia ya existe con información diferente."
              }
            }
          }
        }
      },
      "UnprocessableEntity": {
        "description": "Los IDs existen en el contrato, pero no están disponibles para esta operación.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "required": ["success"],
        "properties": {
          "success": {
            "type": "boolean",
            "const": true
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["success", "error"],
        "properties": {
          "success": {
            "type": "boolean",
            "const": false
          },
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "Pagination": {
        "type": "object",
        "required": ["page", "limit", "total", "total_pages"],
        "properties": {
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "total_pages": {
            "type": "integer"
          }
        }
      },
      "ThirdParty": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "names": {
            "type": "string"
          },
          "lastNames": {
            "type": "string"
          },
          "corporative_name": {
            "type": ["string", "null"],
            "description": "Razón social cuando el registro representa una persona jurídica."
          },
          "full_name": {
            "type": "string"
          },
          "indentification_type": {
            "type": "string"
          },
          "indentification_number": {
            "type": "string"
          },
          "mail": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": ["string", "null"]
          },
          "city": {
            "type": ["string", "null"]
          },
          "address": {
            "type": ["string", "null"]
          },
          "type": {
            "type": "string",
            "enum": ["client", "supplier", "employee", "contractor", "partner", "other", "both"]
          }
        }
      },
...
      "ProcessStep": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": ["string", "null"]
          },
          "order": {
            "type": "integer"
          },
          "optional": {
            "type": "boolean"
          },
          "end_process": {
            "type": "boolean"
          }
        }
      },
...
      "ProcessInstance": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "serial": {
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "start_date": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "delivery_date": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "process": {
            "type": "object"
          },
          "current_step": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ProcessStep"
              },
              {
                "type": "null"
              }
            ]
          },
          "third_party": {
            "type": ["object", "null"]
          },
          "responsible": {
            "type": ["object", "null"]
          }
        }
      },
      "CreateProcessInstanceRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["external_reference", "process_id", "thirdParty_id", "delivery_date"],
        "properties": {
          "external_reference": {
            "$ref": "#/components/schemas/ExternalReference"
          },
          "process_id": {
            "type": "integer",
            "minimum": 1
          },
          "thirdParty_id": {
            "type": "integer",
            "minimum": 1
          },
          "start_date": {
            "type": "string",
            "format": "date-time",
            "description": "Si se omite, el backend utiliza la fecha actual."
          },
          "delivery_date": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string",
            "maxLength": 1000
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "CreateProcessInstanceResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SuccessResponse"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "external_reference": {
                    "type": "string"
                  },
                  "replayed": {
                    "type": "boolean"
                  },
                  "instance_id": {
                    "type": "integer"
                  },
                  "serial": {
                    "type": "integer"
                  },
                  "status": {
                    "type": "string"
                  },
                  "process": {
                    "type": "object"
                  },
                  "current_step": {
                    "$ref": "#/components/schemas/ProcessStep"
                  },
                  "third_party": {
                    "type": "object"
                  },
                  "responsible": {
                    "type": "object"
                  }
                }
              }
            }
          }
        ]
      },
...
      "ExternalReference": {
        "type": "string",
        "minLength": 1,
        "maxLength": 160,
        "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,159}$",
        "description": "Identificador idempotente generado por el sistema externo."
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}